Add files via upload
This Interactive Quiz Application is designed to test your knowledge of HTML, CSS, and JavaScript, three of the most essential technologies for front-end web development. Whether you're a beginner or looking to refresh your knowledge, this quiz provides a fun and challenging way to test your understanding of the basics of web development.